% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssb.R
\name{ssb}
\alias{ssb}
\title{Data from Statics Norway (SSB) Statbank database}
\usage{
ssb(id, lang = "en")
}
\arguments{
\item{id}{Data set ID. See the ssb_search() function.}

\item{lang}{Language of data set to return. Can be "en" for english or "no" for norwegian.}
}
\value{
Data frame with SSB data
}
\description{
Downloads the requested data by using Statistics Norway's API, parses the
resulting JSON file, and returns it as a data frame.
}
\examples{
## View data sets related to population
View(ssb_search("population"))

## Download population by municipalities from 1986 to last year
data <- ssb(26975)
str(data)
View(data)
}
\author{
Mikael Poul Johannesson \email{mikael.johannesson@uib.no}
}

