% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssb_list.R
\name{ssb_list}
\alias{ssb_list}
\title{List of available data sets from Statistics Norway (SSB) API}
\usage{
ssb_list(lang = "en")
}
\arguments{
\item{lang}{Language of available data sets to return. Can be "en" for english or "no" for norwegian.}
}
\value{
Data frame listing all available data sets, with ID, title, tags and URLs.
}
\description{
Download an updated list of available data sets from Statistics Norway (SSB)
API. Returns a data frame.
}
\note{
The \code{ssbR} package ships with a local data frame with infromation on all
    data sets available on 03-01-2016. You can retrieve a new list of available data
    sets using \code{ssb_list()}, and give to to \code{ssb_search()} via the \code{list}
    argument.
}
\examples{
## View all data sets available from the Statistics Norway (SSB) API
View(ssb_list())
}

