% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssb_search.R
\name{ssb_search}
\alias{ssb_search}
\title{Search id, title and tags of available data sets from Statistics Norway (SSB) API}
\usage{
ssb_search(..., lang = "en", list = NULL, update_list = FALSE,
  print_results = TRUE, return_results = TRUE)
}
\arguments{
\item{...}{Character string(s) to use as search terms.}

\item{lang}{Language of available data sets to use. Can be "en" for english or "no" for norwegian.}

\item{list}{List of available data sets retrieved with the \code{ssb_list()} function. If not provided,
the function uses a local list for the search (unless \code{update_list} is set to \code{TRUE}).}

\item{update_list}{If \code{TRUE} the function downloads and use an updated list of available data
sets for the search.}

\item{print_results}{If \code{TRUE} it will print the results in the terminal.}

\item{return_results}{If \code{TRUE} it returns the results as data frame, otherwise it returns \code{NULL}.}
}
\value{
Data frame with id, title, tag and URLs of data sets matching the provided search
terms.
}
\description{
Data frame with id, title, tag and URLs for data sets matching the provided search
terms.
}
\examples{
res <- ssb_search("population")
View(res)
}

